requirements.txt
streamlit>=1.18
torch
torchvision
pillow
numpy
matplotlib